backend/requirements.txt
# ------ FastAPI & ASGI Server ------

fastapi
uvicorn[standard]


# ------ Authentication & Security ------

passlib[bcrypt]
bcrypt
python-jose[cryptography]
python-multipart
email-validator


# ------ Cloud & File Storage ------

cloudinary


# ------ Database (MongoDB) ------

fitz
motor
pymongo
pydantic

# ------ OpenAI & AI Helpers ------

openai
langchain
tiktoken
sentence-transformers


# ------ Vector Databases ------

pinecone
pinecone-client
langchain_openai

# ------ HTML & Text Processing ------

beautifulsoup4
lxml


# ------ Utilities ------

python-dotenv
requests
httpx